requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
boto3==1.26.151
botocore==1.29.151
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.25
click==8.1.3
clickhouse-connect==0.5.25
coloredlogs==15.0.1
dataclasses-json==0.5.7
duckdb==0.8.0
exceptiongroup==1.1.1
faiss-cpu==1.7.4
fastapi==0.95.2
flatbuffers==23.5.26
frozenlist==1.3.3
google-api-core==1.34.0
google-api-python-client==2.88.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.3
google-cloud-speech==2.3.0
google-cloud-texttospeech==2.3.0
google-cloud-translate==2.0.2
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.48.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.2
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
jmespath==1.0.1
langchain==0.0.183
libcst==1.0.0
lz4==4.3.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
peewee==3.16.2
peewee-migrations==0.3.32
posthog==3.0.1
proto-plus==1.22.2
protobuf==3.20.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pydub==0.25.1
PyMuPDF==1.22.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
sympy==1.12
telegram==0.0.1
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zstandard==0.21.0
